Add existing file

+++ backend/api/views.py
@@ -8,9 +8,11 @@ from rest_framework.decorators import action
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
 
+from .filters import RecipeFilter
+from .pagination import LimitPageNumberPagination
 from recipes.models import Favorite, Ingredient, Recipe, ShopCart, Tag
-from users.models import UserFollowing
-from .serializers import (CreateOrUpdateRecipes, IngredientSerializer,
+from users.models import User, UserFollowing
+from .serializers import (CreateOrUpdateRecipes, FollowGetSerializer, IngredientSerializer,
                           RecipesSerializer, TagSerializer,
                           UserFollowersSerializer, UserSerializer)
 from .utils import add_to, delete_from
@@ -36,7 +38,8 @@ class RecipesViewSet(viewsets.ModelViewSet):
     queryset = Recipe.objects.all()
     http_method_names = ['get', 'post', 'patch', 'create', 'delete']
     filter_backends = (DjangoFilterBackend,)
-
+    pagination_class = LimitPageNumberPagination
+    filterset_class = RecipeFilter
     def get_serializer_class(self):
         if self.action == "list" or self.action == "retrieve":
             return RecipesSerializer
@@ -91,42 +94,47 @@ class RecipesViewSet(viewsets.ModelViewSet):
 class CustomUserViewSet(UserViewSet):
     queryset = User.objects.all()
     http_method_names = ['get', 'post', 'delete']
+    pagination_class = LimitPageNumberPagination
 
     @action(
         detail=True,
         methods=["POST", "DELETE"],
         permission_classes=[IsAuthenticated]
     )
-    def subscribe(self, request, pk):
+    def subscribe(self, request, id):
         author = self.get_object()
         if request.method == "POST":
-            user_following, created = UserFollowing.objects.get_or_create(
-                author=author, user=request.user)
+            user_following, created = UserFollowing.objects.get_or_create(author=author, user=request.user)
             serializer = UserFollowersSerializer(user_following)
-            return Response(data=serializer.data,
-                            status=status.HTTP_201_CREATED)
-        UserFollowing.objects.filter(
-            author=author,
-            user=request.user).delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            UserFollowing.objects.filter(author=author, user=request.user).delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
 
     @action(
         methods=['get'],
         detail=False,
         permission_classes=[IsAuthenticated]
     )
+
     def me(self, request):
-        serializer = UserSerializer(request.user)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+      serializer = UserSerializer(request.user,
+              context={'request': request,})
+      return Response(serializer.data, status=status.HTTP_200_OK)
 
     @action(
         detail=False,
         methods=['get'],
-        permission_classes=[IsAuthenticated]
+        permission_classes=[IsAuthenticated],
+        pagination_class=LimitPageNumberPagination
     )
+
+
     def subscriptions(self, request):
-        user_following_query = UserFollowing.objects.filter(
-            user=request.user)
-        serializer = UserFollowersSerializer(
-            user_following_query, many=True)
-        return Response(data=serializer.data, status=status.HTTP_200_OK)
+        user = request.user
+        queryset = User.objects.filter(subscribing__user=user)
+        pages = self.paginate_queryset(queryset)
+        serializer = FollowGetSerializer(pages,
+                                         many=True,
+                                         context={'request': request})
+        return self.get_paginated_response(serializer.data)
